requirements.txt
asgiref==3.7.2
boto3==1.29.6
botocore==1.32.6
click==8.1.7
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.7
django-storages==1.14.2
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==21.2.0
jmespath==1.0.1
numpy==1.26.1
opencv-python-headless
packaging==23.2
Pillow==10.1.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.8.0
urllib3==1.26.18
psycopg2